+++ src/app/eth-extended/data/account/pendingTxsCount/TxPendingCountsReader.ts
@@ -4,8 +4,8 @@ import { ITxPendingCountsResult } from "./ITxPendingCountsResult";
 export class TxPendingCountsReader {
     read(data: any) {
         let pendingResult: ITxPendingCountsResult = {
-            inbound: Number(data["pending"]["inbound"]),
-            outbound: Number(data["pending"]["outbound"])
+            inbound: Number(data["inbound"]),
+            outbound: Number(data["outbound"])
         };
         return pendingResult;
     }