+++ packages/eth-extended/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Update to `pending/v3` deepstream API.
+
 # v3.1.0
 
 - Option to disable USD prices functionality by configuration (`usdPrices` config option).